% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogreModelMain.R
\name{adapt_nondetect}
\alias{adapt_nondetect}
\title{adapts values of single results < detection limit.}
\usage{
adapt_nondetect(x_in, x_nd, factor = 0.5)
}
\arguments{
\item{x_in}{name of input data.frame}

\item{x_nd}{vector with substances < detection (one column per site)}

\item{factor}{multiplier of detection limit if smaller dl (e.g., 0, 0.5 or 1)}
}
\description{
requires list of these substances as data.frame for each site (one column per
site). If substance is always < dl at one site, results are set to zero. If
substance is sometimes > dl, resluts are set to a factor*dl
}
