% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogreModelMain.R
\name{annual_load_rain}
\alias{annual_load_rain}
\title{calculates the load for each substance}
\usage{
annual_load_rain(data.dir, error_removal_rate = 0.3)
}
\arguments{
\item{data.dir}{path of model data (annual mean concentrations
"annual_mean_conc.csv",  rain runoff volumes "Vol_rain.csv", removal at
WWTP "substance_info.csv")}

\item{error_removal_rate}{relative error in removal at WWTP}
}
\value{
Function returns list with loads and standard deviations,
by entry path (sep, cso, wwtp) and by surface water catchment.
Concentration in units "mg/L" and "ug/L" is automatically
transformed to loads in "kg/yr". Other (unknown) units are
left unchanged, resulting in "unit * m3/yr".
}
\description{
separates pathways (rain runoff, CSO and WWTP)
}
