% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogreModelMain.R
\name{annual_mean_conc}
\alias{annual_mean_conc}
\title{estimates annual mean concentrations per site.}
\usage{
annual_mean_conc(x_in, method, data.dir)
}
\arguments{
\item{x_in}{name of input data.frame}

\item{method}{estimation method: 1 = arithmetic mean; 2 = functions and RMSE
from file, arithmetic mean for substances without functions}

\item{data.dir}{file directory where correlation data and rain series for
method 2 are located}
}
\description{
Apart from a matrix with mean concentrations for each substance and site (= 0
if always below detection limit), matrices with N, standard error, standard
deviation, RMSE, as well as measured min and max are calculated. Result is
given as a list, of these matrices. Different methods can be chosen:
 Method 1: arithmetic mean
 Method 2: functions and RMSE from file, arithmetic mean for substances
 without functions
}
