% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogreModelMain.R
\name{annual_stats}
\alias{annual_stats}
\title{estimates annual mean concentrations per site.}
\usage{
annual_stats(x_in)
}
\arguments{
\item{x_in}{name of input data.frame}
}
\description{
Apart from a matrix with mean concentrations for each substance and site (= 0
if always below detection limit), matrices with N, 95% confidence, as well as
measured min and max are calculated. Result is given as a list, of these
matrices.
}
