% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogreModelMain.R
\name{default_statistics}
\alias{default_statistics}
\title{calculate default statistics}
\usage{
default_statistics(x)
}
\arguments{
\item{x}{data frame with columns \code{DataValue}, \code{CensorCode}}
}
\description{
calculate default statistics for a grouped data frame (created with
dplyr::group_by)
}
