% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogreModelMain.R
\name{detect}
\alias{detect}
\title{removes substances, without detection from data.frame}
\usage{
detect(x_in, x_nd)
}
\arguments{
\item{x_in}{name of input data.frame}

\item{x_nd}{vector with substances < detection}
}
\description{
requires list of these substances as single vector
}
