% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogreModelMain.R
\name{geom_mean}
\alias{geom_mean}
\title{function gives geometric mean}
\usage{
geom_mean(x)
}
\arguments{
\item{x}{vector of numeric values of which to calculate the geometric mean}
}
\description{
function gives geometric mean
}
