% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogreModelMain.R
\name{get_lab_values}
\alias{get_lab_values}
\title{Reads all lab values from ODBC-source}
\usage{
get_lab_values(odbc_name)
}
\arguments{
\item{odbc_name}{Name of the odbc source}
}
\description{
Appends also site code (e.g., "NEU") and substance name apart from
standard-fields the fields "CensorCode", "QualityControlLevelID", and
"UnitsAbbreviation" are included.  If one measurements exists twice, only
higher QualityControlLevelID is included
}
