% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogreModelMain.R
\name{no_Panke}
\alias{no_Panke}
\title{removes rows in data.frame with site code = "PNK"}
\usage{
no_Panke(x_in)
}
\arguments{
\item{x_in}{name of input data.frame}
}
\description{
removes rows in data.frame with site code = "PNK"
}
