% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogreModelMain.R
\name{non_detect}
\alias{non_detect}
\title{lists substances without detection in any sample}
\usage{
non_detect(x_in)
}
\arguments{
\item{x_in}{name of input data.frame}
}
\description{
Apart from entire dataset x_in (first column), lists are given for each site
individually (following columns)
}
