% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogreModelMain.R
\name{only_composite}
\alias{only_composite}
\title{removes rows in data.frame with SampleType ! = "Composite"}
\usage{
only_composite(x_in)
}
\arguments{
\item{x_in}{name of input data.frame}
}
\description{
removes rows in data.frame with SampleType ! = "Composite"
}
