% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogreModelMain.R
\name{only_new_dl_metals}
\alias{only_new_dl_metals}
\title{removes metal samples below detection limit (dl),}
\usage{
only_new_dl_metals(x_in)
}
\arguments{
\item{x_in}{name of input data.frame}
}
\description{
when dl was too high (old analytical method). Works by VariableCode
}
