% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogreModelMain.R
\name{only_representative_subst}
\alias{only_representative_subst}
\title{removes Variables, which have not}
\usage{
only_representative_subst(x_in)
}
\arguments{
\item{x_in}{name of input data.frame}
}
\description{
at least one measurement (can also be below detection limit) per catchment
type
}
