% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogreModelMain.R
\name{Panke}
\alias{Panke}
\title{keeps only rows in data.frame with site code = "PNK"}
\usage{
Panke(x_in)
}
\arguments{
\item{x_in}{name of input data.frame}
}
\description{
keeps only rows in data.frame with site code = "PNK"
}
