% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogreModelMain.R
\name{quant25}
\alias{quant25}
\title{function gives the 25 percent quantile}
\usage{
quant25(x)
}
\arguments{
\item{x}{vector of numeric values of which to calculate the quantile}
}
\description{
function gives the 25 percent quantile
}
