% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogreModelMain.R
\name{quant75}
\alias{quant75}
\title{function gives the 75 percent quantile}
\usage{
quant75(x)
}
\arguments{
\item{x}{vector of numeric values of which to calculate the quantile}
}
\description{
function gives the 75 percent quantile
}
