% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogreModelMain.R
\name{quant95}
\alias{quant95}
\title{function gives the 95 percent quantile}
\usage{
quant95(x)
}
\arguments{
\item{x}{vector of numeric values of which to calculate the quantile}
}
\description{
function gives the 95 percent quantile
}
