% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogreModelMain.R
\name{reduce_codes}
\alias{reduce_codes}
\title{removes lines with censor codes}
\usage{
reduce_codes(x_in)
}
\arguments{
\item{x_in}{name of input data.frame}
}
\description{
other than "lt" and "nc" (e.g., "???" are removed)
}
