% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogreModelMain.R
\name{remove_group}
\alias{remove_group}
\title{removes measurements of Variables of a specific group}
\usage{
remove_group(x_in, group)
}
\arguments{
\item{x_in}{name of input data.frame}

\item{group}{Variable group to be removed (string)}
}
\description{
removes measurements of Variables of a specific group
}
